Allow queries to complete before closing TSM files
If a query was running against a file being compacted, we close the file
and the query would end wherever it had read up to.  This could result
in queries that randomly lost data, but running them again showed the
full results.

We now use a reference counting approach and move the in-use files out
of the way in the filestore and allow the queries to complete against
the old tsm files.  The new files are installed and new queries will
use them.

To move the files out of the way, the existing file is hardlink and
remap with the existing TSMReader which is quick as we do not need
to reload the index (which is very slow).  This should also work
cross platform better than just renaming the old file (which breaks
on windows if the file is still open).

Fixes #5501